Update patch set 2
Patch Set 2:

> Yes I think having the Jenkins job in place first is the right
 > strategy also.  I've given that one +1 and added the required LF
 > reviewers.  Now we wait.

Chris,

I'm opposed to removing the mock username, passwords, and component URLs because that's essential information for reproducing the example.

Regarding license, I plan on taking care of that and other cosmetic changes once we're out of crunch mode for DCAE integration. Likewise, this is a mechanical task, so if it's very important then others are free to PR the change.

As for the release date, I haven't previously included dates. Isn't that what VCSs are for? Can you point to an authority on required release note structure?

Regarding my name and email, why would I want to remove that information?

And a .keep file keeps packages/ so that the example stays consistent.

Patch-set: 2
Label: Code-Review=0