Update patch set 5
Patch Set 5:

> Paul - thanks for your work. Please ensure that your commit message
 > adheres to the Commit Message format guidelines https://wiki.acumos.org/display/AC/Acumos+CI-CD+Resources+and+Processes
 >  Also see this patch for an example of a good commit message
 > format: https://gerrit.acumos.org/r/#/c/524/
 > 
 > Licenses: Every file that supports comments MUST have a
 > license/copyright header. https://wiki.acumos.org/display/AC/Licensing
 >  has guidelines as well as license headers for a number of
 > languages. Contact Bryan Sullivan in Research if you need a script
 > for license insertion.
 > 
 > Docs: All documentation should be in RST as the primary format.
 > Please remove the MD files and the gen_rst script. Also, the docs
 > folder needs a developers-guide.rst as well as a users-guide.rst.
 > Reach out to Lorraine Welch in Research for help/guidance as
 > needed.

Aimee, can you please provide a link to the RST authority document? I.e. where is this mandated? I have seen other repositories with markdown files. Also, what is the policy on language-specific markup files, for example Roxygen and the R language?

I personally think markdown is a much more readable format, and that its prevalence  in open source projects is an indicator of that. If this is a subject matter, then I would like to maintain markdown files on the grounds of readability and productivity.

Patch-set: 5
Label: Code-Review=0